+++ source/keystone_osk/__init__.py
@@ -5,4 +5,4 @@
 
 from __future__ import annotations
 
-__version__ = "0.9.0"
+__version__ = "0.9.1"